Create playRound function
The function takes the playerSelection and computerSelection
parameters, makes both selections lowercase, and compares them
to return a result of a single round. The result will either
declare a tie or a win or loss for the user. Currently the user
input is hardcoded so a prompt will need to be added.